#342 [chore] 알림 센터 관련 패키지 정리
