+++ app/src/main/java/com/spark/android/data/remote/datasource/AlarmCenterDataSourceImpl.kt
@@ -6,7 +6,7 @@ import androidx.paging.PagingData
 import com.spark.android.data.remote.entity.response.Alarm
 import com.spark.android.data.remote.entity.response.NoDataResponse
 import com.spark.android.data.remote.service.AlarmCenterService
-import com.spark.android.ui.alarmcenter.acitivityalarm.pagingsource.AlarmPagingSource
+import com.spark.android.ui.alarmcenter.alarms.pagingsource.AlarmPagingSource
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
