+++ app/src/main/java/com/spark/android/ui/alarmcenter/alarms/adapter/AlarmPagingAdapter.kt
@@ -1,4 +1,4 @@
-package com.spark.android.ui.alarmcenter.acitivityalarm.adapter
+package com.spark.android.ui.alarmcenter.alarms.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
@@ -8,7 +8,7 @@ import androidx.recyclerview.widget.RecyclerView
 import com.spark.android.BR
 import com.spark.android.data.remote.entity.response.Alarm
 import com.spark.android.databinding.ItemAlarmListBinding
-import com.spark.android.ui.alarmcenter.acitivityalarm.adapter.AlarmPagingAdapter.ActivityAlarmViewHolder
+import com.spark.android.ui.alarmcenter.alarms.adapter.AlarmPagingAdapter.ActivityAlarmViewHolder
 
 class AlarmPagingAdapter :
     PagingDataAdapter<Alarm, ActivityAlarmViewHolder>(activityAlarmDiffUtil) {
