+++ app/src/main/res/layout/fragment_activity_alarm.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.alarmcenter.acitivityalarm.ActivityAlarmFragment">
+        tools:context=".ui.alarmcenter.alarms.ActivityAlarmFragment">
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/rv_activity"
